Add missing period/duration parsing & formatting.
